fix(post): syncPosts에서 삭제 대상 없을 때 예외 처리 추가

+++ src/post/post.service.ts
@@ -86,8 +86,11 @@ export class PostService {
       .filter((p) => !posts.includes(p.id))
       .map((p) => p.id);
 
-    const res = await this.postRepository.delete(deletedPostsIds);
-    console.log(res);
+    if (deletedPostsIds.length != 0) {
+      const res = await this.postRepository.delete(deletedPostsIds);
+
+      console.log(res);
+    }
 
     return {
       success: true,
